+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 11.2
+---------------
+- Add $add_post argument, POST parameters to add to the URL (optional) in Prepare_PHP_SELF.fnc.php
+
 Changes in 11.1.2
 -----------------
 - Maintain list preferences as GET params in form URL (save) in Schedule.php